Refactor quantum state handling and enhance normalization processes
- Updated the `QuantumState` class to convert amplitudes and phases to `complex64` instead of `complex128`, improving memory efficiency.
- Modified normalization steps in the `_to_quantum_state` and `_from_quantum_state` methods to ensure global normalization of states, enhancing consistency in quantum state conversions.
- Improved tensor handling in the `QuantumGeometricTensor` class for parameter derivatives, ensuring compatibility with complex tensors.
- Enhanced test cases in `test_quantum.py` to validate quantum state conversions and ensure proper normalization, including detailed tensor information logging for better debugging.

These changes improve the robustness and efficiency of quantum state management, aligning with best practices for numerical analysis and testing.

+++ src/core/quantum/types.py
@@ -20,16 +20,16 @@ class QuantumState:
 
     def __post_init__(self):
         """Ensure state normalization and proper tensor types."""
-        # Convert to complex128 if not already
+        # Convert to complex64 if not already
         if not torch.is_complex(self.amplitudes):
-            self.amplitudes = self.amplitudes.to(torch.complex128)
-        elif self.amplitudes.dtype != torch.complex128:
-            self.amplitudes = self.amplitudes.to(torch.complex128)
+            self.amplitudes = self.amplitudes.to(torch.complex64)
+        elif self.amplitudes.dtype != torch.complex64:
+            self.amplitudes = self.amplitudes.to(torch.complex64)
             
         if not torch.is_complex(self.phase):
-            self.phase = self.phase.to(torch.complex128)
-        elif self.phase.dtype != torch.complex128:
-            self.phase = self.phase.to(torch.complex128)
+            self.phase = self.phase.to(torch.complex64)
+        elif self.phase.dtype != torch.complex64:
+            self.phase = self.phase.to(torch.complex64)
 
         # Normalize state vector
         if len(self.amplitudes.shape) == 1:
@@ -43,9 +43,9 @@ class QuantumState:
 
         # Store original norm if not provided
         if self.original_norm is None:
-            self.original_norm = norm.to(torch.float64)
+            self.original_norm = norm.to(torch.float32)
         else:
-            self.original_norm = self.original_norm.to(torch.float64)
+            self.original_norm = self.original_norm.to(torch.float32)
 
     @property
     def shape(self) -> Tuple[int, ...]:
@@ -60,8 +60,8 @@ class QuantumState:
     def norm(self) -> torch.Tensor:
         """Compute the norm of the quantum state."""
         if len(self.amplitudes.shape) == 1:
-            return torch.sqrt(torch.sum(torch.abs(self.amplitudes) ** 2)).to(torch.float64)
-        return torch.sqrt(torch.sum(torch.abs(self.amplitudes) ** 2, dim=1)).to(torch.float64)
+            return torch.sqrt(torch.sum(torch.abs(self.amplitudes) ** 2)).to(torch.float32)
+        return torch.sqrt(torch.sum(torch.abs(self.amplitudes) ** 2, dim=1)).to(torch.float32)
 
     def density_matrix(self) -> torch.Tensor:
         """Compute the density matrix representation of the state."""
